more error handling

+++ src/files_api/routes.py
@@ -102,6 +102,15 @@ async def get_file_metadata(
     Note: by convention, HEAD requests MUST NOT return a body in the response.
     """
     s3_bucket_name = request.app.state.settings.s3_bucket_name
+
+    object_exists = object_exists_in_s3(
+        bucket_name=s3_bucket_name, object_key=file_path
+    )
+    if not object_exists:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="File not found"
+        )
+
     get_object_response = fetch_s3_object(s3_bucket_name, object_key=file_path)
     response.headers["Content-Type"] = get_object_response["ContentType"]
     response.headers["Content-Length"] = str(get_object_response["ContentLength"])
@@ -144,6 +153,14 @@ async def delete_file(
 
     NOTE: DELETE requests MUST NOT return a body in the response."""
     s3_bucket_name = request.app.state.settings.s3_bucket_name
+
+    object_exists = object_exists_in_s3(
+        bucket_name=s3_bucket_name, object_key=file_path
+    )
+    if not object_exists:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="File not found"
+        )
     delete_s3_object(s3_bucket_name, object_key=file_path)
     response.status_code = status.HTTP_204_NO_CONTENT
     return response
